Feature/refactor

+++ life_expectancy/tests/conftest.py
@@ -10,7 +10,7 @@ def run_before_and_after_tests() -> None:
     """Fixture to execute commands before and after a test is run"""
     # Setup: fill with any logic you want
 
-    yield # this is where the testing happens
+    yield  # this is where the testing happens
 
     # Teardown : fill with any logic you want
     file_path = OUTPUT_DIR / "pt_life_expectancy.csv"
@@ -24,7 +24,7 @@ def pt_life_expectancy_expected() -> pd.DataFrame:
         FIXTURES_DIR / "pt_life_expectancy_expected.csv"
     )
     data_expected["value"] = data_expected[
-        "value"].astype("float32")
+        "value"].astype("float64")
     return data_expected
